backend/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.10.0
Werkzeug==2.3.7
alembic==1.12.1 
psycopg2==2.9.6
python-dateutil==2.9.0.post0
